cleaned up sample data

+++ lib/tasks/sample_data.rake
@@ -1,9 +1,29 @@
+require 'rubygems'
+require 'faker'
+
+def colorize(text, color_code)
+ "\e[#{color_code}m#{text}\e[0m"
+end
+
+def red(text);      colorize(text, 31); end
+def green(text);    colorize(text, 32); end
+def yellow(text);   colorize(text, 33); end
+def purple(text);   colorize(text, 34); end
+def magenta(text);  colorize(text, 35); end
+def cyan(text);     colorize(text, 36); end
+
 namespace :db do
   desc "Fill database with sample data" 
   task populate: :environment do
-    Rake::Task['db:reset'].invoke
+     if Rails.env.development?
+      puts "#{red("==>")} Clearing Current Data"
+      Rake::Task['db:reset'].invoke
+    end
+    puts "#{green("==>")} Creating sample admin user"
     sample_admin
+    puts "#{green("==>")} Creating sample users"
     make_users
+    puts "#{green("==>")} Creating merciboqs"
     make_merciboqs
   end
 end
@@ -20,7 +40,7 @@ end
 def make_users
   99.times do |n|
     name = Faker::Name.name 
-    email = "user#{n+1}@merciboq.com"
+    email = Faker::Internet.email
     password  = "password"
     user = User.new(name: name,
                     :email => email,
